requirements.txt
boto3==1.16.40
botocore==1.19.40
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
DateTime==4.3
gitdb==4.0.5
github3.py==0.9.3
GitPython==3.1.11
idna==2.10
jmespath==0.10.0
jwcrypto==0.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
smmap==3.0.4
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
